+++ src/main/java/org/scion/multiping/Main.java
@@ -22,16 +22,19 @@ import org.scion.multiping.util.Util;
 public class Main {
 
   public static void main(String[] args) throws IOException {
-    if (args.length != 1) {
-      printUsage();
-      System.exit(1);
-    }
+    checkArgs(args, 1, Integer.MAX_VALUE);
     String mode = args[0].toLowerCase(Locale.ROOT);
     String[] newArgs = Arrays.copyOfRange(args, 1, args.length);
     switch (mode) {
       case "download-assignments":
         {
+          checkArgs(args, 1, 1);
           DownloadAssignments.main(newArgs);
+          return;
+        }
+      case "help":
+        {
+          printHelp(args.length == 1 ? "" : args[1]);
           return;
         }
       case "ping-all":
@@ -41,17 +44,46 @@ public class Main {
         }
       case "ping-repeat":
         {
+          checkArgs(args, 1, 1);
           PingRepeat.main(newArgs);
           return;
         }
       case "ping-responder":
         {
+          checkArgs(args, 1, 1);
           PingResponder.main(newArgs);
           return;
         }
       default:
         printUsage();
         System.exit(1);
+    }
+  }
+
+  private static void checkArgs(String[] args, int minArgs, int maxArgs) {
+    if (args.length < minArgs || args.length > maxArgs) {
+      Util.println("Invalid number of arguments.");
+      printUsage();
+      System.exit(1);
+    }
+  }
+
+  private static void printHelp(String mode) {
+    switch (mode) {
+      case "download-assignments":
+        printUsageDownloadAssignments();
+        return;
+      case "ping-all":
+        printUsagePingAll();
+        return;
+      case "ping-repeat":
+        printUsagePingRepeat();
+        return;
+      case "ping-responder":
+        printUsagePingResponder();
+        return;
+      default:
+        printUsage();
     }
   }
 
@@ -65,6 +97,53 @@ public class Main {
         "    - `ping-repeat` for repeatedly probing (traceroute) multiple paths to multiple ASes.");
     Util.println(
         "    - `ping-responder` for starting a server that responds to incoming echo requests.");
+    Util.println("    - `help [MODE]` for getting more help for a given mode.");
+    Util.println("");
+  }
+
+  private static void printUsageDownloadAssignments() {
+    Util.println("Usage: scion-multiping download-assignments");
+    Util.println();
+    Util.println(
+        "  This tool downloads a list of known ISD/AS assignments and saves it to a file.");
+    Util.println("  The output file is called `isd-as-assignments.csv`.");
+    Util.println("");
+  }
+
+  static void printUsagePingAll() {
+    Util.println(
+        "Usage: ping-all [--help] [--fastest|--shortest|--shortest_echo|--fastest_sync] [--port <port>] [--shim]");
+    Util.println("  --help              Show this help message.");
+    Util.println("  --fastest           Use fastest path with SCMP traceroute (default).");
+    Util.println(
+        "                      The fastest path is determined by running a single traceroute on all path.");
+    // Util.println("  --fastest_sync      Use fastest path with SCMP traceroute (synchronous)");
+    Util.println("  --shortest          Use shortest path (fewest hops) with SCMP traceroute.");
+    // Util.println("  --shortest_echo     Use shortest path with SCMP echo");
+    Util.println(
+        "  --port <port>       Use specified local port (default " + PingAll.localPort + ").");
+    Util.println("  --shim              Start with SHIM enabled (default disabled).");
+    Util.println("");
+  }
+
+  private static void printUsagePingRepeat() {
+    Util.println("Usage: scion-multiping ping-repeat");
+    Util.println();
+    Util.println(
+        "  This tool is used for repeatedly probing (traceroute) multiple paths to multiple ASes.");
+    Util.println("  The destination ASes are read from a file `isd-as-assignments.csv`.");
+    Util.println("  Other configuration options can be defined in `ping-repeat-config.json`.");
+    Util.println("  Results are written to a CSV file `ping-results.csv`.");
+    Util.println("  See README.md for more information.");
+    Util.println("");
+  }
+
+  private static void printUsagePingResponder() {
+    Util.println("Usage: scion-multiping ping-responder");
+    Util.println();
+    Util.println("  This command starts a server that responds to incoming echo requests.");
+    Util.println("  It takes a configuration file `ping-responder-config.json` as input.");
+    Util.println("  See README.md for more information.");
     Util.println("");
   }
 }